TST: GH26807 separate out test_retain_attributes
